Replacing setState conditionals with conditional rendering.
Reason: While conditionally calling setState is functional, it obscures
the rationale behind the conditions.  The desired effect is actually a
gated `render` call, which is better accomplished by the
`shouldComponentUpdate` lifecycle method.  This simplifies the
`recomputeState` method while making the effect more explicit.

As a side-effect of this change, it was easier to see that changes to
`this.state.height` should only trigger renders while sticky.