Replace context-based updates with a Channel.
Reason: As stated in the previous commit, context is not a stable way
to communicate data updates between non-adjacent components – it is,
however, a reasonable way to pass data *once*.  Using this information,
we can establish a link between the parent Container and child that
becomes a pub/sub communications channel.

This also allows us to do away with Container state, yielding a
performance boost (since `setState` would trigger unnecessary
re-renders of the Container and its children).